added a bumnch of tests
